pyproject.toml
[project]
name = "drive-ocr"
version = "0.1.0"
description = "Google Drive document analysis using ChatGPT"
requires-python = ">=3.13"

dependencies = [
  "google-api-python-client>=2.120.0",
  "openai>=1.3.5",
  "pydantic>=2.6.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project.scripts]
drive-ocr = "drive_ocr.main:main"
